Rename KeepIntervalForService to...
...CalculateAdjectiveForService
