+++ lib/expire.rb
@@ -12,9 +12,10 @@ require 'expire/calculate_service_base'
 require 'expire/calculate_adjective_service_base'
 require 'expire/calculate_adjective_service'
 require 'expire/calculate_adjective_for_service_base'
+require 'expire/calculate_adjective_for_service'
 require 'expire/calculate_service'
 require 'expire/keep_service_base'
-require 'expire/keep_interval_for_service'
+# require 'expire/keep_interval_for_service'
 require 'expire/keep_interval_for_from_now_service'
 require 'expire/result'
 require 'expire/rules'
