+++ lib/expire/calculate_adjective_for_service.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Expire
+  # Calculate expiration until
+  class CalculateAdjectiveForService < CalculateAdjectiveForServiceBase
+    def now
+      @now ||= backups.max
+    end
+
+    def rule
+      "#{adjective}_for"
+    end
+  end
+end
