+++ spec/expire/calculate_adjective_for_service_spec.rb
@@ -4,7 +4,7 @@ require 'test_dates'
 require 'support/shared_examples_for_keep_until_services'
 require 'support/shared_examples_for_keep_services'
 
-RSpec.describe Expire::KeepIntervalForService do
+RSpec.describe Expire::CalculateAdjectiveForService do
   it_behaves_like 'a keep service' do
     let(:constructor_args) {{}} # { { now: :fake_now } }
   end
